requirements.txt
ECHO is on.
Flask==2.3.2
python-docx==0.8.11
reportlab==3.6.9
Werkzeug==2.3.3
gunicorn==20.1.1

#pip install -r requirements.txt
